Cleanup - remove account command - simplify root command - remove service command - replace service dev command - update process command based on root command
