+++ packages/cli/src/commands/process/delete.ts
@@ -1,4 +1,4 @@
-import {flags} from '@oclif/command'
+import { flags } from '@oclif/command'
 import cli from 'cli-ux'
 
 import Command from '../../root-command'
@@ -9,7 +9,7 @@ export default class ProcessDelete extends Command {
 
   static flags = {
     ...Command.flags,
-    confirm: flags.boolean({description: 'Confirm deletion', default: false}),
+    confirm: flags.boolean({ description: 'Confirm deletion', default: false }),
     account: flags.string({
       description: 'Account to use to delete the process'
     })
@@ -23,16 +23,16 @@ export default class ProcessDelete extends Command {
   }]
 
   async run(): Promise<string[]> {
-    const {argv, flags} = this.parse(ProcessDelete)
+    const { argv, flags } = this.parse(ProcessDelete)
     const { account, mnemonic } = await this.getAccount(flags.account)
     if (!flags.confirm && !await cli.confirm('Are you sure?')) return []
-    this.spinner.start('Deleting process(s)')
+    cli.action.start('Deleting process(s)')
     const tx = await this.lcd.createTransaction(
       argv.map(hash => this.lcd.process.deleteMsg(account.address, hash)),
       account
     )
     const txResult = await this.lcd.broadcast(tx.signWithMnemonic(mnemonic))
-    this.spinner.stop()
+    cli.action.stop()
     return findHash(txResult, 'process', 'DeleteProcess')
   }
 }
