+++ packages/cli/src/commands/process/dev.ts
@@ -1,14 +1,15 @@
-import {IProcess} from '@mesg/api/lib/process'
-import {hash} from '@mesg/api/lib/types'
+import { IProcess } from '@mesg/api/lib/process'
+import { hash } from '@mesg/api/lib/types'
 import * as base58 from '@mesg/api/lib/util/base58'
 
 import Command from '../../root-command'
-import {IsAlreadyExistsError} from '../../utils/error'
+import { IsAlreadyExistsError } from '../../utils/error'
 
 import ProcessCompile from './compile'
 import ProcessCreate from './create'
 import ProcessDelete from './delete'
 import ProcessLog from './logs'
+import { cli } from 'cli-ux'
 
 export default class ProcessDev extends Command {
   static description = 'Run a process in development mode'
@@ -29,17 +30,17 @@ export default class ProcessDev extends Command {
   instanceCreated = false
 
   async run() {
-    const {args, flags} = this.parse(ProcessDev)
+    const { args, flags } = this.parse(ProcessDev)
 
-    this.spinner.start('Starting process')
-    this.spinner.status = 'compiling'
+    cli.action.start('Starting process')
+    cli.action.status = 'compiling'
     const envToArgs = (flags.env || []).reduce((prev: string[], next: string) => [...prev, '--env', next], [])
-    const definition = await ProcessCompile.run([args.PROCESS, '--silent', '--dev', ...envToArgs, ...this.flagsAsArgs(flags)])
-    this.spinner.status = 'creating process'
+    const definition = await ProcessCompile.run([args.PROCESS, '--dev', ...envToArgs, ...this.flagsAsArgs(flags)])
+    cli.action.status = 'creating process'
     const processHash = await this.createProcess(definition)
-    this.spinner.status = 'fetching logs'
+    cli.action.status = 'fetching logs'
     const stream = await ProcessLog.run([base58.encode(processHash), ...this.flagsAsArgs(flags)])
-    this.spinner.stop('ready')
+    cli.action.stop('ready')
 
     process.once('SIGINT', async () => {
       stream.destroy()
